Klar

+++ Programmering 2/Receptprogrammet, övning 4.4/Receptprogrammet, övning 4.4/Ingridiens.cs
@@ -9,9 +9,9 @@ namespace Receptprogrammet__övning_4._4
     internal class Ingridiens
     {
         public string Namn { get; set; }
-        public int Mängd { get; set; }
+        public float Mängd { get; set; }
         public string Mått { get; set; }
-        public Ingridiens(string _namn, int _mängd, string _mått)
+        public Ingridiens(string _namn, float _mängd, string _mått)
         {
             Namn = _namn;
             Mängd = _mängd;
